Handled errors when sex-count is initially 0

+++ relMan/sex/views.py
@@ -4,6 +4,8 @@ from django.shortcuts import render, redirect
 import datetime
 from .models import Sex
 from django.contrib.auth.decorators import login_required
+from django.shortcuts import get_object_or_404
+
 
 # Create your views here.
 
@@ -15,7 +17,11 @@ def index(request):
 @login_required
 def sex_view(request):
     sex_count = Sex.objects.filter(sex_user=request.user).count()
-    last_sex = Sex.objects.latest('created')
+    check_sex = Sex.objects.all()
+    if check_sex.exists():
+        last_sex = Sex.objects.latest('created')
+    else:
+        last_sex = 0
 
     context = {
         'sex_count': sex_count,
@@ -35,7 +41,10 @@ def addSex(request):
 
 @login_required
 def deleteSex(request):
-    sex = Sex.objects.filter(sex_user=request.user).latest('created')
-
-    sex.delete()
+    # sex = Sex.objects.filter(sex_user=request.user).latest('created')
+    if Sex.objects.all().exists():
+        sex = Sex.objects.filter(sex_user=request.user).latest('created')
+        sex.delete()
+    else:
+        sex = 0
     return redirect('/sex-counter')
